% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fda_bspline}
\alias{fda_bspline}
\title{Spline-Basis Preprocessing Spec}
\usage{
fda_bspline(
  df = 6L,
  degree = 3L,
  intercept = TRUE,
  center = FALSE,
  scale = FALSE
)
}
\arguments{
\item{df}{Degrees of freedom used by \code{\link[splines:bs]{splines::bs()}}.}

\item{degree}{Spline degree.}

\item{intercept}{Should the spline basis include an intercept column?}

\item{center, scale}{Logical flags controlling column-wise centering and
scaling of the resulting coefficients.}
}
\value{
An object of class \code{fda_preprocess_spec}.
}
\description{
Spline-Basis Preprocessing Spec
}
