% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fda_identity}
\alias{fda_identity}
\title{Identity Preprocessing Spec}
\usage{
fda_identity(center = FALSE, scale = FALSE)
}
\arguments{
\item{center, scale}{Logical flags controlling column-wise centering and
scaling of the transformed features.}
}
\value{
An object of class \code{fda_preprocess_spec}.
}
\description{
Identity Preprocessing Spec
}
