% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fda_standardize}
\alias{fda_standardize}
\title{Standardization Preprocessing Spec}
\usage{
fda_standardize(center = TRUE, scale = TRUE)
}
\arguments{
\item{center, scale}{Logical flags controlling column-wise centering and
scaling. Both default to \code{TRUE}.}
}
\value{
An object of class \code{fda_preprocess_spec}.
}
\description{
Standardization Preprocessing Spec
}
