% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{fit_selectboost_formula}
\alias{fit_selectboost_formula}
\title{Fit FDA-SelectBoost from a Formula}
\usage{
fit_selectboost_formula(
  formula,
  data,
  family = c("gaussian", "binomial"),
  transforms = NULL,
  scalar_transform = NULL,
  preprocessor = NULL,
  center = FALSE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{formula, data, family, transforms, scalar_transform, preprocessor, center, scale}{Passed to \code{\link[=fda_design_formula]{fda_design_formula()}}.}

\item{...}{Additional arguments passed to \code{\link[=fit_selectboost]{fit_selectboost()}}.}
}
\value{
An object inheriting from \code{selectboost_fda_result}.
}
\description{
Fit FDA-SelectBoost from a Formula
}
