% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit_stability}
\alias{fit_stability}
\title{Fit Grouped Stability Selection from an FDA Design}
\usage{
fit_stability(design, ...)
}
\arguments{
\item{design}{An \code{fda_design} object.}

\item{...}{Additional arguments forwarded to \code{stability_selection_fda()}.}
}
\value{
An object inheriting from \code{fda_stability_selection}.
}
\description{
Fit Grouped Stability Selection from an FDA Design
}
