% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{fit_stability_formula}
\alias{fit_stability_formula}
\title{Fit Stability Selection from a Formula}
\usage{
fit_stability_formula(
  formula,
  data,
  family = c("gaussian", "binomial"),
  transforms = NULL,
  scalar_transform = NULL,
  preprocessor = NULL,
  center = FALSE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{formula, data, family, transforms, scalar_transform, preprocessor, center, scale}{Passed to \code{\link[=fda_design_formula]{fda_design_formula()}}.}

\item{...}{Additional arguments passed to \code{\link[=fit_stability]{fit_stability()}}.}
}
\value{
An object inheriting from \code{fda_stability_selection}.
}
\description{
Fit Stability Selection from a Formula
}
