% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-extractors.R
\name{report_formula_blocks}
\alias{report_formula_blocks}
\title{Report Formula Blocks}
\usage{
report_formula_blocks()
}
\value{
A named list of character strings.
}
\description{
Returns named LaTeX formula strings used in report material.
}
