% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation-grid.R
\name{summary.fda_perturbation_grid}
\alias{summary.fda_perturbation_grid}
\title{Summarize a Perturbation Grid}
\usage{
\method{summary}{fda_perturbation_grid}(object, ...)
}
\arguments{
\item{object}{A \code{fda_perturbation_grid} object.}

\item{...}{Unused.}
}
\value{
A summary object.
}
\description{
Summarize a Perturbation Grid
}
