% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit_selectboost}
\alias{fit_selectboost}
\title{Fit SelectBoost from an FDA Design}
\usage{
fit_selectboost(design, ...)
}
\arguments{
\item{design}{An \code{fda_design} object.}

\item{...}{Additional arguments forwarded to \code{selectboost_fda()}.}
}
\value{
An object inheriting from \code{selectboost_fda_result}.
}
\description{
Fit SelectBoost from an FDA Design
}
